Support custom content scope indicator for DAM
