+++ packages/admin/cms-admin/src/dam/DamPage.tsx
@@ -1,16 +1,49 @@
+import { messages } from "@comet/admin";
+import { Domain } from "@comet/admin-icons";
+import { Typography } from "@mui/material";
+import { styled } from "@mui/material/styles";
 import * as React from "react";
+import { FormattedMessage } from "react-intl";
 
+import { ContentScopeIndicator } from "../contentScope/ContentScopeIndicator";
+import { ContentScopeInterface, useContentScope } from "../contentScope/Provider";
 import { useContentScopeConfig } from "../contentScope/useContentScopeConfig";
 import { DamTable } from "./DamTable";
 
 type Props = {
     path: string;
+    renderContentScopeIndicator?: (scope: ContentScopeInterface) => React.ReactNode;
 };
 
-function DamPage({ path }: Props): React.ReactElement {
+const defaultRenderContentScopeIndicator = () => (
+    <ContentScopeIndicator variant="toolbar" global>
+        <ScopeIndicatorContent>
+            <Domain fontSize="small" />
+            <ScopeIndicatorLabelBold variant="body2">
+                <FormattedMessage {...messages.globalContentScope} />
+            </ScopeIndicatorLabelBold>
+        </ScopeIndicatorContent>
+    </ContentScopeIndicator>
+);
+
+const ScopeIndicatorLabelBold = styled(Typography)`
+    && {
+        font-weight: 400;
+        padding: 0 8px 0 4px;
+        text-transform: uppercase;
+    }
+`;
+
+const ScopeIndicatorContent = styled("div")`
+    display: flex;
+    align-items: center;
+`;
+
+function DamPage({ path, renderContentScopeIndicator = defaultRenderContentScopeIndicator }: Props): React.ReactElement {
+    const { scope } = useContentScope();
     useContentScopeConfig({ redirectPathAfterChange: path });
 
-    return <DamTable />;
+    return <DamTable contentScopeIndicator={renderContentScopeIndicator(scope)} />;
 }
 
 export { DamPage };
